Revert "Corrección de color y de vista de perfil"
