+++ src/App.js
@@ -4,7 +4,6 @@ import './styles/App.css';
 import axios from 'axios';
 
 import Login from './components/login/login';
-import Perfil from './components/perfil/perfil';
 import Home from './components/home/home';
 import Signup from './components/signup/signup';
 import Editar_perfil from './components/editar_perfil/editar_perfil';
@@ -28,7 +27,6 @@ const App = () => {
           <Route path="/login" element={<Login />}></Route>
           <Route path="/restablecer_contraseña" element={<Restablecer_contraseña />}></Route>
           <Route path="/home" element={<Home />}></Route>
-          <Route path="/perfil" element={<Perfil />}></Route>
         </Routes>
       </BrowserRouter>
