+++ src/components/header_post/header_post.js
@@ -3,7 +3,6 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import Logo from '../img/cropped-logo_positivo.png';
 import NavDropdown from 'react-bootstrap/NavDropdown';
 import React from 'react';
-import { Link } from 'react-router-dom';
 
 class Header_post extends React.Component {
 
@@ -21,7 +20,7 @@ class Header_post extends React.Component {
                         aria-expanded="false"
                         aria-label="Toggle navigation"
                     >
-                        <FontAwesomeIcon className="color-icon-l" icon={faBars} />
+                        <FontAwesomeIcon icon={faBars} />
                     </button>
 
                     <div className="collapse navbar-collapse" id="navbarSupportedContent">
@@ -35,16 +34,12 @@ class Header_post extends React.Component {
                         </a>
                     </div>
                     <div>
-                        <Link to="/home">
-                            <button
-                                className="btn btn-link ms-3 text-dark"
-                                type="button"
-                            >
-                                <FontAwesomeIcon className="color-icon-l" icon={faHouse} size="xl" />
-
-                            </button>
-                        </Link>
-
+                        <button
+                            className="btn btn-link ms-3 text-dark"
+                            type="button"
+                        >
+                            <FontAwesomeIcon icon={faHouse} size="xl" />
+                        </button>
                     </div>
 
                     <div class="input-group rounded ms-4 me-5">
@@ -58,7 +53,7 @@ class Header_post extends React.Component {
                     <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
                         <li class="nav-item me-2">
                             <div className="d-flex align-items-center">
-                                <FontAwesomeIcon className="color-icon-l" icon={faUser} size="xl" />
+                                <FontAwesomeIcon icon={faUser} size="xl" />
                                 <NavDropdown id="navbarScrollingDropdown">
                                     <NavDropdown.Item href="/login">Iniciar sesión</NavDropdown.Item>
                                     <NavDropdown.Item href="/register">Registrarse</NavDropdown.Item>
@@ -67,7 +62,7 @@ class Header_post extends React.Component {
                         </li>
                         <li class="nav-item">
                             <div className="d-flex align-items-center space">
-                                <FontAwesomeIcon className="color-icon-l" icon={faUniversalAccess} size="xl" />
+                                <FontAwesomeIcon icon={faUniversalAccess} size="xl" />
                                 <NavDropdown id="navbarScrollingDropdown">
                                     <NavDropdown.Item href="#action4"> Ajuste del nivel de contraste</NavDropdown.Item>
                                     <NavDropdown.Item href="#action4"> Ajuste del tamaño de la letra</NavDropdown.Item>
