+++ src/styles/header.css
@@ -8,7 +8,4 @@
     border-bottom: 2px solid #006186;
  }
  
- .color-icon-l {
-   color: #006186 !important;
-}
